requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9